Updating News Service API
The previous api integrated into the application was a large fail, as I did not understand the requirments for
the api pricing. Only on localhost was the api free, otherwise a charge was introduced.

A new weather API was integrated into the application that allows 100 requests per day for free.
This allows the small unvisited application a chance to shine without breaking the wallet.

A few more style and data display updates were made to the application as well