Version 1 The first official version of Weather App React has been released!!
In this application a user can view the weather from their area via location that is automatically detected
via permissons in the browser.

The data then loads and a display for Todays weather, hourly weather in a 24 hour period, and daily weather for
the week is displayed on a clean looking interface.

This application also features tool tips, top 3 news articles, day and night styling
based on your time of day and the sunset/sunrise data given from open meteo API,
fully mobile responsive & hourly notifications in the browser